Refactor: DetailPage 스켈레톤 컴포넌트화 및 폴더 구조 변경 (#1)
