+++ src/components/detail/InfoSection/InfoSectionSkeleton.tsx
