+++ src/components/detail/index.tsx
@@ -1,23 +1,26 @@
 import SubMenuBar from "./SubMenuBar/SubMenuBar";
+import SubMenuBarSkeleton from "./SubMenuBar/SubMenuBarSkeleton";
 import SubMenuSection from "./SubMenuSection/SubMenuSection";
 import LikeButton from "./LikeButton/LikeButton";
 import InfoSection from "./InfoSection/InfoSection";
-import InfoSectionSkeleton from "./InfoSectionSkeleton/InfoSectionSkeleton";
+import InfoSectionSkeleton from "./InfoSection/InfoSectionSkeleton";
 import ReviewSection from "./ReviewSection/ReviewSection";
-import ReviewSectionSkeleton from "./ReviewSectionSkeleton/ReviewSectionSkeleton";
+import ReviewSectionSkeleton from "./ReviewSection/ReviewSectionSkeleton";
 import ReviewForm from "./ReviewForm/ReviewForm";
 import ReviewItem from "./ReviewItem/ReviewItem";
 import ReviewEditForm from "./ReviewEditForm/ReviewEditForm";
 import ReservationModal from "./ReservationModal/ReservationModal";
 import ReservationForm from "./ReservationForm/ReservationForm";
 import ReservationPayment from "./ReservationPayment/ReservationPayment";
-import { DetailPageSkeleton, ButtonGroupSkeleton, BannerSkeleton } from "./DetailPageSkeleton/DetailPageSkeleton";
 import Banner from "./Banner/Banner";
+import BannerSkeleton from "./Banner/BannerSkeleton";
 import ButtonGroup from "./ButtonGroup/ButtonGroup";
+import ButtonGroupSkeleton from "./ButtonGroup/ButtonGroupSkeleton";
 
 export {
-  SubMenuSection,
   SubMenuBar,
+  SubMenuBarSkeleton,
+  SubMenuSection,
   LikeButton,
   InfoSection,
   InfoSectionSkeleton,
@@ -33,5 +36,4 @@ export {
   BannerSkeleton,
   ButtonGroup,
   ButtonGroupSkeleton,
-  DetailPageSkeleton,
 };
